Moved original class BaseController classes to _core Added wrapper to original BaseController
